added mysql workbench dependency and database and changed user credential handling across the board
