+++ src/main/resources/application.properties
@@ -1,3 +1,7 @@
 server.port = 9085
 
 spring.jackson.serialization.indent_output = true
+
+spring.datasource.url=jdbc:mysql://localhost:3306/redcom
+spring.datasource.username=root
+spring.datasource.password=DamV@nDam00
